# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieApp",
        "description": "Api para vizualização de informações de filmes migrados a partir de uma dataset",
        "contact": {
            "name": "Vitor Guedes",
            "email": "vtrf2.0@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://dev.backend.com",
            "description": "Ambiente local",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http",
                    "description": "Protocolo usado"
                }
            }
        }
    ],
    "paths": {
...
        "/v1/api/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Lista de reviews do usuário",
                "operationId": "1662a9b920a45b4812b2a2b1706c0470",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Cria uma nova review do usuário para o filme",
                "operationId": "aa3ae6f12311e84b8e540dc2404f5ed4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "review": {
                                        "description": "Texto contendo o conteúdo da review",
                                        "type": "string"
                                    },
                                    "movie_id": {
                                        "description": "Id do filme que vai receber a review",
                                        "type": "integer"
                                    },
                                    "positive": {
                                        "description": "Se a review é positiva ou negativa",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Se foi um sucesso",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Texto de retorno da requisição",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/v1/api/reviews/{reviewId}": {
            "put": {
                "tags": [
                    "Review"
                ],
                "summary": "Atualiza uma review do usuário para o filme",
                "operationId": "7c88b390b7961832f19bd068aac9deab",
                "parameters": [
                    {
                        "name": "reviewId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "review": {
                                        "description": "Texto contendo o conteúdo da review",
                                        "type": "string"
                                    },
                                    "positive": {
                                        "description": "Se a review é positiva ou negativa",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Se foi um sucesso",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Texto de retorno da requisição",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "400": {
                        "description": "Parametros inválidos"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Review"
                ],
                "summary": "Deleta uma review do usuário para o filme",
                "operationId": "2f6a4e5c54ddd37c583c2475be6685cb",
                "parameters": [
                    {
                        "name": "reviewId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Se foi um sucesso",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Texto de retorno da requisição",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "400": {
                        "description": "Parametros inválidos"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "description": "Token de acesso",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}